+++ R/Maze.R
@@ -48,7 +48,8 @@
 #' gw <- gridworld_init(dim = c(3, 4), unreachable_states = c("s(2,2)"))
 #' gridworld_matrix(gw)
 #'
-#' # the transition function is stochastic so we cannot use gw$T()
+#' # the transition function is stochastic so we cannot use the standard
+#' # gridworld gw$transition_prob() function
 #' T <- function(action, start.state, end.state) {
 #'   action <- match.arg(action, choices = gw$actions)
 #'